+++ partiql-spi/src/test/kotlin/org/partiql/spi/value/DatumStringTest.kt
@@ -1,0 +1,43 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+package org.partiql.spi.value
+
+import org.junit.jupiter.api.Test
+import kotlin.test.assertEquals
+
+class DatumStringTest {
+    @Test
+    fun varcharPreservesValuesShorterThanLength() {
+        val datum = Datum.varchar("a", 20)
+
+        assertEquals("a", datum.string)
+    }
+
+    @Test
+    fun varcharPreservesValuesAtLength() {
+        val value = "a".repeat(20)
+        val datum = Datum.varchar(value, 20)
+
+        assertEquals(value, datum.string)
+    }
+
+    @Test
+    fun varcharTruncatesValuesLongerThanLength() {
+        val datum = Datum.varchar("a".repeat(30), 20)
+
+        assertEquals("a".repeat(20), datum.string)
+    }
+}